Revert "some warnings on windows x64"
This reverts commit 14bc2c865da6042f37e334e965c714c77bcabde2.

+++ scilab/modules/parallel/src/cpp/parallel_run.hxx
@@ -303,13 +303,13 @@ namespace
             if (dynamic_scheduling)
             {
 #pragma omp parallel for private(i) schedule(dynamic, chunk_size)
-                for(i=0; (size_t)i < n; ++i)
+                for(i=0; i < n; ++i)
                 {
                     callF(i);
                 }
             } else {
 #pragma omp parallel for private(i)  schedule(static, chunk_size)
-                for(i=0; (size_t)i < n; ++i)
+                for(i=0; i < n; ++i)
                 {
                     callF(i);
                 }
